fix bugs in preprocess

+++ datas/raw_dataset.py
@@ -128,10 +128,10 @@ class RawShangHang(Dataset):
         self.data_path = os.path.join(root, data_filename)
         raw = sio.loadmat(self.data_path)
         if split == 'train':
-            self.data = raw['DataCube1'].astype('float32')
+            self.data = raw['DataCube1'].astype('float16')
             self.gt = raw['gt1'].astype('int')
         else:
-            self.data = raw['DataCube2'].astype('float32')
+            self.data = raw['DataCube2'].astype('float16')
             self.gt = raw['gt2'].astype('int')
 
         self.transform = transform
